+++ CHANGELOG.rst
@@ -2,9 +2,19 @@
 Changelog
 =========
 
-Version 0.1
+Version 0.1.x
 ===========
 
-- Feature A added
-- FIX: nasty bug #1729 fixed
-- add your changes here!
+The goal of this version is to provide the basic templates to handle DSP Modules in the form of a class with a configuration step and a processing step. It is a general framework that can always work in any projects, eg. in a streaming context. Its configuration parameters are safe and are a fit for automated GUI generation.
+
+- Parameters [add] : parameter template for DSPModule
+- Parameters [add] : float parameter
+
+TODO
+===========
+
+- Parameters [evo] : opt arg to parameters to handle a list of accepted values
+- Parameters [add] : string parameter
+- Parameters [add] : int parameter
+- Modules [add] : DSPModule template
+- Modules [add] : Simple gain DSPModule